+++ src/test/java/io/github/davidhlp/spring/cache/redis/cache/ActualCacheHandlerIntegrationTest.java
@@ -60,7 +60,7 @@ class ActualCacheHandlerIntegrationTest extends AbstractRedisIntegrationTest {
     private ValueOperations<String, Object> valueOperations;
 
     @Autowired
-    private NullValuePolicy nullValuePolicy;
+    private NullValueEncoder nullValueEncoder;
 
     @Mock
     private RefreshCancellation earlyExpirationExecutor;
@@ -77,7 +77,7 @@ class ActualCacheHandlerIntegrationTest extends AbstractRedisIntegrationTest {
         handler = new ActualCacheHandler(
                 redisTemplate,
                 valueOperations,
-                nullValuePolicy,
+                nullValueEncoder,
                 earlyExpirationExecutor,
                 errorHandler);
     }
@@ -111,7 +111,7 @@ class ActualCacheHandlerIntegrationTest extends AbstractRedisIntegrationTest {
         if (onDelete != null) {
             when(throwingTpl.delete("test:key")).thenThrow(onDelete);
         }
-        return new ActualCacheHandler(throwingTpl, throwingOps, nullValuePolicy,
+        return new ActualCacheHandler(throwingTpl, throwingOps, nullValueEncoder,
                 earlyExpirationExecutor, errorHandler);
     }